EMG Detection Bound on Servo Motor
Implemented if statement that will compare input pin of EMG sensor to see if it is producing an output. If output is detected from the output pin, the motor will be allowed to run forward, else, the natural tension and friction of elastic band will allow the string to fall back to neutral.